Implement Retain Email Log setting

+++ uninstall.php
@@ -34,14 +34,21 @@ if ( is_multisite() ) {
 function email_log_delete_table() {
 	global $wpdb;
 
+	$retain_email_log = get_option( 'el_email_log' );
+
 	// This is hardcoded on purpose, since the entire plugin is not loaded during uninstall.
 	$table_name = $wpdb->prefix . 'email_log';
 
 	if ( $wpdb->get_var( "SHOW TABLES LIKE '{$table_name}'" ) == $table_name ) {
-		// If table is present, drop it
-		$wpdb->query( "DROP TABLE $table_name" );
+		if ( ! empty( $retain_email_log ) &&
+		     is_array( $retain_email_log ) &&
+		     ! array_key_exists( 'retain_email_logs', $retain_email_log ) ) {
+			// If table is present, drop it
+			$wpdb->query( "DROP TABLE $table_name" );
+		}
 	}
 
-	// Delete the option
+	// Delete the options.
 	delete_option( 'email-log-db' );
+	delete_option( 'el_email_log' );
 }
